% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea}
\alias{dea}
\title{Data envelopment analysis}
\usage{
dea(XREF, YREF, X, Y, model = "output", RTS = "variable")
}
\arguments{
\item{XREF}{Matrix of inputs for observations used for constructing the frontier}

\item{YREF}{Matrix of outputs for observations used for constructing the frontier}

\item{X}{Matrix of inputs for observations, for which DEA scores are estimated}

\item{Y}{Matrix of outputs for observations, for which DEA scores are estimated}

\item{model}{Orientation of the model; must be "input" or "output"}

\item{RTS}{Returns-to-scale for the model; must be "constant", "non-increasing" or "variable"}
}
\value{
Returns a list with the following components
\item{thetaOpt}{A vector of efficiency estimates, in [0, 1] interval}
\item{lambda}{A matrix of constraint coefficients}
\item{lambda_sum}{A vector of sum of lambdas; lambda_sum = 1 for variable RTS, lambda_sum <= for non-increasing RTS}
\item{model}{Orientation of the model}
\item{RTS}{Returns-to-scale for the model}
}
\description{
Basic data envelopment analysis to replace rDEA::dea due to difficulties
installing rDEA on travis-ci
}
\details{
This function estimates efficiency using data envelopment analysis. The linear program
is constructed as in Fare et al. (1985) and optimized using lpSolve::lp. The function
was built to swap with rDEA::dea because installation of rDEA presented problems for
travis-ci, due to difficulties installing glpk. Should installation of rDEA on travis-ci
become possible, this function will be removed.
}
\examples{
data(univariate)

dea.fit <- dea(univariate$x, univariate$y,
               univariate$x, univariate$y,
               model = "output",
               RTS = "variable")
univariate$frontier <- univariate$y / dea.fit$thetaOpt

#Plot technical/allocative efficiency over time
library(ggplot2)

ggplot(univariate, aes(x, y)) +
  geom_point() +
  geom_line(aes(y = frontier), color = "red")

}
\references{
\insertRef{FareDEA}{snfa}
}
