% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy}
\alias{energy}
\title{Energy sources and sinks}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
energy
}
\description{
Dataset describing energy generation and consumption
as a directed network. Data come from the Department
of Energy & Climate Change via Tom Counsell. See
\url{http://www.decc.gov.uk/en/content/cms/tackling/2050/calculator_on/calculator_on.aspx}.
}
\keyword{datasets}
