% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey}
\alias{sankey}
\title{Sankey diagram}
\usage{
sankey(
  df,
  source.column = "source",
  target.column = "target",
  value.column = "value",
  align = c("justify", "left", "right", "center"),
  edge.color = c("path", "input", "output", "none"),
  width = NULL,
  height = NULL,
  viewer = c("internal", "external", "browser")
)
}
\arguments{
\item{df}{data.frame containing edge data}

\item{source.column}{Name of column containing source nodes. Defaults to "source".}

\item{target.column}{Name of column containing target nodes. Defaults to "target".}

\item{value.column}{Name of column containing edge values. Defaults to "value".}

\item{align}{Alignment of node labels. Defaults to "justify".}

\item{edge.color}{Method of coloring edges. The value "path" will create a gradient between two nodes. Defaults to "path".}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}

\item{viewer}{"internal" to use the RStudio internal viewer pane for output; "external" to display in an external RStudio window; "browser" to display in an external browser.}
}
\value{
A d3 object as returned by r2d3::r2d3.
}
\description{
Creates Sankey diagram from edge data.frame.
}
\details{
Utilizes a script similar to \url{https://observablehq.com/@d3/sankey-diagram} adapted to work with r2d3.
}
\examples{
data(energy)

sankey(energy)

}
