% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.d3.R
\name{save.d3}
\alias{save.d3}
\title{Save d3 diagram as png}
\usage{
save.d3(
  d3,
  file,
  width = 1000,
  height = 750,
  delay = 5,
  zoom = 1,
  background = "white",
  title = "D3 Visualization"
)
}
\arguments{
\item{d3}{A d3 object.}

\item{file}{Location to save image.}

\item{width}{Width of image.}

\item{height}{Height of image.}

\item{zoom}{Zoom before screenshot.}

\item{background}{Background color of diagram.}

\item{title}{Title for HTML diagram.}
}
\description{
Saves d3 diagram as a png image using webshot.
}
\examples{
data(energy)

d3 = sankey(energy)
save.d3(d3, "energy.png")

}
